Refactor
Reduce Clunkiness of App.py
Introduce Blackboard Session Manager
