+++ .gitignore
@@ -11,3 +11,4 @@ backend/credentials.json
 .DS_Store
 frontend/.DS_Store
 backend/support/.DS_Store
+frontend/.DS_Store
